requirements.txt
ace_tools==0.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
arxiv==2.2.0
asttokens==3.0.0
asyncio==3.4.3
attrs==23.2.0
blobfile==3.0.0
build==1.2.2.post1
cbor==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
disutils==1.4.32.post2
docutils==0.21.2
executing==2.2.0
feedparser==6.0.11
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gymnasium>=1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
id==1.5.0
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.7
llama_models==0.0.63
llama_stack==0.0.63
llama_stack_client==0.0.63
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
msgpack>=1.0.0
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.4.1
nh3==0.2.21
numpy==2.1.2
openai==1.76.2
packaging==24.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.0
pyproject_hooks==1.2.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
safetensors==0.4.5
seaborn==0.13.2
setuptools==75.2.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sse-starlette==2.2.1
stack-data==0.6.3
stardojo==0.1
starlette==0.46.2
sympy==1.13.1
termcolor==2.5.0
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.21.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.51.3
twine==6.1.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.2
wcwidth==0.2.13
yarl==1.9.4
